doc/cephadm: clarify ssh key flags
This PR clairfies that one should pass file paths to the --ssh-private-key
and --ssh-public-key bootstrap flags.

The current manpage and helptext for cephadm isn't sufficiently clear on
whether I should pass the path to the key file(s) or the actual key
itself. Intuitively, it's pretty obvious that it should be a file path
for a *lot* of reasons, but I still like to be sure before I do things,
and this required me to go to the source code to see how the flags are
used.

Signed-off-by: thomas <me@thomasanderson.cloud>